+++ ff4j-core/src/main/java/org/ff4j/store/AbstractFeatureStore.java
@@ -22,6 +22,7 @@ package org.ff4j.store;
 
 import java.io.InputStream;
 import java.util.Collection;
+import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
 
@@ -110,6 +111,86 @@ public abstract class AbstractFeatureStore implements FeatureStore {
         sb.append("]");
         sb.append("}");
         return sb.toString();
+    }
+
+    /** {@inheritDoc} */
+    @Override
+    public void enableGroup(String groupName) {
+        Map < String, Feature > features = readGroup(groupName);
+        for (Map.Entry<String,Feature> uid : features.entrySet()) {
+            uid.getValue().enable();
+            update(uid.getValue());
+        }
+    }
+
+    /** {@inheritDoc} */
+    @Override
+    public void disableGroup(String groupName) {
+        Map < String, Feature > features = readGroup(groupName);
+        for (Map.Entry<String,Feature> uid : features.entrySet()) {
+            uid.getValue().disable();
+            update(uid.getValue());
+        }
+    }
+
+    /** {@inheritDoc} */
+    @Override
+    public void addToGroup(String featureId, String groupName) {
+        Util.assertParamHasLength(groupName, "groupName (#2)");
+        // retrieve
+        Feature f = read(featureId);
+        f.setGroup(groupName);
+        // persist modification
+        update(f);
+    }
+
+    /** {@inheritDoc} */
+    @Override
+    public void removeFromGroup(String featureId, String groupName) {
+        Util.assertParamHasLength(groupName, "groupName (#2)");
+        if (!existGroup(groupName)) {
+            throw new GroupNotFoundException(groupName);
+        }
+        // retrieve
+        Feature f = read(featureId);
+        f.setGroup(null);
+        // persist modification
+        update(f);
+    }
+
+    /** {@inheritDoc} */
+    @Override
+    public Set<String> readAllGroups() {
+        Map < String, Feature > features = readAll();
+        Set < String > groups = new HashSet<String>();
+        for (Map.Entry<String,Feature> uid : features.entrySet()) {
+            groups.add(uid.getValue().getGroup());
+        }
+        groups.remove(null);
+        return groups;
+    }
+
+    /** {@inheritDoc} */
+    @Override
+    public void grantRoleOnFeature(String flipId, String roleName) {
+        Util.assertParamHasLength(roleName, "roleName (#2)");
+        // retrieve
+        Feature f = read(flipId);
+        // modify
+        f.getPermissions().add(roleName);
+        // persist modification
+        update(f);
+    }
+
+    /** {@inheritDoc} */
+    @Override
+    public void removeRoleFromFeature(String flipId, String roleName) {
+        Util.assertParamHasLength(roleName, "roleName (#2)");
+        // retrieve
+        Feature f = read(flipId);
+        f.getPermissions().remove(roleName);
+        // persist modification
+        update(f);
     }
 
     /** {@inheritDoc} */
